Derailed Capabilities
- more porting
- port JEI plugin
- use new capability lifecycle for minecarts

+++ src/main/java/com/simibubi/create/content/curiosities/weapons/PotatoProjectileRenderMode.java
@@ -34,7 +34,7 @@ public interface PotatoProjectileRenderMode {
 
 			MatrixTransformStack.of(ms)
 				.rotateY(AngleHelper.deg(Mth.atan2(diff.x, diff.z)) + 180)
-				.rotateX(AngleHelper.deg(Mth.atan2(diff.y, Mth.sqrt(diff.x * diff.x + diff.z * diff.z))));
+				.rotateX(AngleHelper.deg(Mth.atan2(diff.y, Mth.sqrt((float) (diff.x * diff.x + diff.z * diff.z)))));
 		}
 
 	}
@@ -71,7 +71,7 @@ public interface PotatoProjectileRenderMode {
 			MatrixTransformStack.of(ms)
 				.rotateY(AngleHelper.deg(Mth.atan2(diff.x, diff.z)))
 				.rotateX(270
-					+ AngleHelper.deg(Mth.atan2(diff.y, -Mth.sqrt(diff.x * diff.x + diff.z * diff.z))));
+					+ AngleHelper.deg(Mth.atan2(diff.y, -Mth.sqrt((float) (diff.x * diff.x + diff.z * diff.z)))));
 			MatrixTransformStack.of(ms)
 				.rotateY((entity.tickCount + pt) * 20 * spin + entityRandom(entity, 360))
 				.rotateZ(-spriteAngleOffset);
